Add DatabaseHelper & Product classes
The first iteration of the DatabaseHelper class. It includes the functions 'addOne' and 'selectAll'. addOne takes a Product object and inserts it into the database, and selectAll returns an ArrayList of all the items currently in the Product table.

I included some sample code from my MainActivity.java which I was using to test out the DatabaseHelper class.

Note: right now, app will crash if you try to insert a blank record.