Add Spinner DatePicker to AddFragment
Got rid of the masked input text box and replaced it with a Spinner-style DatePicker dialog.
NOTE: If the app is crashing when you try to run it on your phone, try going to App info --> Storage and press "Clear data".
I am not sure why this is happening, maybe the old database sticks around even after you uninstall the app? But this fixed it for me.