Fix Category insertion bug
Database schema change:
- added a UNIQUE constraint onto category name to prevent duplicates.
Bugfixes:
There was two bugs in the Add Fragment. The first one was that an a duplicate value was getting added to the 'categories' ArrayList. That would cause the insert product query to sometimes fail if you had recently deleted a category, since the duplicate value would still exist in the program but not in the database, so SQLite would throw an error if you tried to use a nonexistant categoryID as a foreign key. This would result in the category getting inserted but the product not.
Long story short, it needed a transaction around both queries, so if either of them fail, they both get rolled back. The implementation is more of a quickfix though.